Added Controller classes
Added controller and changed context classes

+++ BNTFileSystemProgram/BNTFileSystemProgram/DataLayer/AuthorContext.cs
@@ -61,12 +61,29 @@ namespace DataLayer
             }
         }
 
+        public async Task UpdateAsync(string key)
+        {
+            try
+            {
+                Author? author = await _context.Authors.FindAsync(key);
+                if (author != null)
+                {
+                    _context.Authors.Update(author);
+                    await _context.SaveChangesAsync();
+                }
+            }
+            catch (Exception ex)
+            {
+
+                throw new Exception(ex.Message);
+            }
+        }
         public async Task UpdateAsync(Author item)
         {
             try
             {
-                _context.Authors.Update(item);
-                await _context.SaveChangesAsync();
+                    _context.Authors.Update(item);
+                    await _context.SaveChangesAsync();
             }
             catch (Exception ex)
             {
@@ -80,8 +97,11 @@ namespace DataLayer
             try
             {
                 var authorFromDb = await _context.Authors.FindAsync(key);
-                _context.Authors.Remove(authorFromDb);
-                await _context.SaveChangesAsync();
+                if (authorFromDb != null)
+                {
+                    _context.Authors.Remove(authorFromDb);
+                    await _context.SaveChangesAsync();
+                }
             }
             catch (Exception ex)
             {
